refactor(HttpRequest): HttpRequest parsing 객체 분리
