+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/HttpRequest.java
@@ -1,10 +1,7 @@
 package org.apache.coyote.http11.message.request;
 
-import java.io.BufferedReader;
-import java.io.IOException;
 import java.net.URI;
 import java.util.Map;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.coyote.http11.HttpCookies;
 import org.apache.coyote.http11.message.common.ContentType;
 import org.apache.coyote.http11.message.common.HttpBody;
@@ -14,8 +11,6 @@ import org.apache.util.parser.Parser;
 
 public class HttpRequest {
 
-    private static final int OFFSET = 0;
-
     private final HttpRequestLine startLine;
     private final HttpHeaders headers;
     private final HttpBody body;
@@ -24,33 +19,6 @@ public class HttpRequest {
         this.startLine = startLine;
         this.headers = headers;
         this.body = body;
-    }
-
-    public static HttpRequest from(BufferedReader reader) throws IOException {
-        HttpRequestLine httpRequestLine = new HttpRequestLine(parseStartLine(reader));
-        HttpHeaders httpHeaders = new HttpHeaders(parseHeaders(reader));
-        HttpBody httpBody = new HttpBody(parseBody(reader, httpHeaders.getContentLength()));
-
-        return new HttpRequest(httpRequestLine, httpHeaders, httpBody);
-    }
-
-    private static String parseStartLine(BufferedReader reader) throws IOException {
-        return reader.readLine();
-    }
-
-    private static String parseHeaders(BufferedReader reader) throws IOException {
-        StringBuilder builder = new StringBuilder();
-        String line;
-        while (StringUtils.isNoneBlank(line = reader.readLine())) {
-            builder.append(line).append(System.lineSeparator());
-        }
-        return builder.toString();
-    }
-
-    private static String parseBody(BufferedReader reader, int countLength) throws IOException {
-        char[] buffer = new char[countLength];
-        reader.read(buffer, OFFSET, countLength);
-        return new String(buffer);
     }
 
     public Map<String, String> getKeyValueBodies() {
